+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/Models/Responses/TypeSpec/TypeSpecBaseResponse.cs
@@ -3,13 +3,17 @@ using Azure.Sdk.Tools.Cli.Attributes;
 
 namespace Azure.Sdk.Tools.Cli.Models.Responses.TypeSpec
 {
-    public class TypeSpecBaseResponse
+    /// <summary>
+    /// Serves as a base class for TypeSpec-related responses, providing common telemetry properties.
+    /// </summary>
+    public abstract class TypeSpecBaseResponse : CommandResponse
     {
         [Telemetry]
         [JsonPropertyName("typespec_project")]
-        public string? TypeSpecProject { get; set; }
+        [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
+        public required string TypeSpecProject { get; set; } = string.Empty;
         [JsonPropertyName("package_type")]
         [Telemetry]
         public SdkType PackageType { get; set; }
-        }
+    }
 }